refactor(reporting): simplificar exportação com rota direta, renderers dedicados e tokens do DESIGN.md
- Remove rota assíncrona, tasks em background e upload desnecessário no Storage para relatórios sob demanda.
- Extrai render_wedding_pdf e render_wedding_excel para pdf_utils e excel_utils aplicando a paleta oficial (#7C3AED Prestige Purple).
- Cria wedding_report_data_selector e DTO imutável reduzindo duplicação no ReportGenerationService.
- Limpa o frontend com dropdown direto (PDF e Excel) e feedback visual de loading com Sonner toasts.
- Atualiza testes automatizados no backend e frontend com 100% de sucesso.
