+++ backend/apps/reporting/api.py
@@ -11,7 +11,6 @@ from pydantic import UUID4
 from apps.core.constants import READ_ERROR_RESPONSES
 from apps.reporting.schemas import (
     DashboardSummaryOut,
-    ReportTaskQueuedOut,
     WeddingDashboardOut,
 )
 from apps.reporting.selectors import (
@@ -19,9 +18,7 @@ from apps.reporting.selectors import (
     wedding_overview_selector,
 )
 from apps.reporting.services import ReportGenerationService
-from apps.reporting.tasks import generate_wedding_report_task
 from apps.users.types import AuthRequest
-from apps.weddings.selectors import wedding_get_selector
 
 
 dashboard_router = Router(tags=["Dashboard"])
@@ -77,70 +74,16 @@ def export_wedding_report(
     format: Literal["pdf", "excel"] = "pdf",
 ) -> HttpResponse:
     """
-    Gera e exporta em fluxo binário síncrono o relatório consolidado do casamento.
+    Gera e exporta em fluxo binário o relatório consolidado do casamento.
 
     Retorna o arquivo binário com Content-Disposition correspondente ao formato.
     """
-    user = request.user
-    company = user.company
-
-    if format == "excel":
-        file_bytes = ReportGenerationService.generate_wedding_excel(
-            company=company,
-            wedding_uuid=uuid,
-        )
-        content_type = (
-            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
-        )
-        filename = f"relatorio-casamento-{uuid}.xlsx"
-    else:
-        file_bytes = ReportGenerationService.generate_wedding_pdf(
-            company=company,
-            wedding_uuid=uuid,
-        )
-        content_type = "application/pdf"
-        filename = f"relatorio-casamento-{uuid}.pdf"
+    file_bytes, content_type, filename = ReportGenerationService.export_wedding_report(
+        company=request.user.company,
+        wedding_uuid=uuid,
+        report_format=format,
+    )
 
     response = HttpResponse(file_bytes, content_type=content_type)
     response["Content-Disposition"] = f'attachment; filename="{filename}"'
     return response
-
-
-@reports_router.post(
-    "/weddings/{uuid}/async/",
-    response={202: ReportTaskQueuedOut, **READ_ERROR_RESPONSES},
-    operation_id="reports_wedding_export_async",
-)
-def export_wedding_report_async(
-    request: AuthRequest,
-    uuid: UUID4,
-    format: Literal["pdf", "excel"] = "pdf",
-) -> tuple[int, dict[str, str]]:
-    """
-    Dispara a geração de relatório em background task (django.tasks).
-
-    O arquivo é processado pelo worker, salvo no Cloudflare R2 / S3 e uma
-    notificação in-app é enviada ao usuário com o link seguro de download.
-    """
-    user = request.user
-    company = user.company
-
-    # Validação prévia de isolamento multi-tenant
-    wedding_get_selector(company=company, uuid=uuid)
-
-    generate_wedding_report_task.enqueue(
-        company_id=str(company.uuid),
-        user_id=str(user.uuid),
-        wedding_id=str(uuid),
-        report_format=format,
-    )
-
-    fmt_label = format.upper()
-    detail_msg = (
-        f"Geração do relatório ({fmt_label}) iniciada em segundo plano. "
-        "Você receberá uma notificação quando estiver pronto."
-    )
-    return 202, {
-        "status": "enqueued",
-        "detail": detail_msg,
-    }
