+++ backend/apps/reporting/selectors/__init__.py
@@ -2,6 +2,10 @@ from .dashboard_selectors import (
     dashboard_summary_selector,
     wedding_overview_selector,
 )
+from .report_selectors import (
+    WeddingReportDataDTO,
+    wedding_report_data_selector,
+)
 from .summaries import (
     ContractSummarySelector,
     FinancialSummarySelector,
@@ -13,6 +17,8 @@ __all__ = [
     "ContractSummarySelector",
     "FinancialSummarySelector",
     "TaskSummarySelector",
+    "WeddingReportDataDTO",
     "dashboard_summary_selector",
     "wedding_overview_selector",
+    "wedding_report_data_selector",
 ]
