+++ frontend/src/features/reporting/components/ExportReportDropdown.test.tsx
@@ -17,7 +17,7 @@ describe("ExportReportDropdown", () => {
 
     const button = screen.getByTestId("export-report-dropdown-trigger");
     expect(button).toBeInTheDocument();
-    expect(button).toHaveTextContent("Exportar Relatório");
+    expect(button).toHaveTextContent("Exportar");
   });
 
   it("opens the dropdown menu on click displaying PDF and Excel options", async () => {
@@ -27,15 +27,11 @@ describe("ExportReportDropdown", () => {
     const button = screen.getByTestId("export-report-dropdown-trigger");
     await user.click(button);
 
-    expect(screen.getByText("Download Imediato")).toBeInTheDocument();
-    expect(screen.getByTestId("export-pdf-sync")).toBeInTheDocument();
-    expect(screen.getByTestId("export-excel-sync")).toBeInTheDocument();
-    expect(screen.getByText("Segundo Plano (Worker)")).toBeInTheDocument();
-    expect(screen.getByTestId("export-pdf-async")).toBeInTheDocument();
-    expect(screen.getByTestId("export-excel-async")).toBeInTheDocument();
+    expect(screen.getByTestId("export-pdf")).toBeInTheDocument();
+    expect(screen.getByTestId("export-excel")).toBeInTheDocument();
   });
 
-  it("triggers synchronous PDF export on click", async () => {
+  it("triggers PDF export on click", async () => {
     server.use(
       http.get("*/api/v1/reports/weddings/:uuid/", () => {
         return new HttpResponse("%PDF-1.4 Mock Binary", {
@@ -58,7 +54,7 @@ describe("ExportReportDropdown", () => {
     const button = screen.getByTestId("export-report-dropdown-trigger");
     await user.click(button);
 
-    const pdfOption = screen.getByTestId("export-pdf-sync");
+    const pdfOption = screen.getByTestId("export-pdf");
     await user.click(pdfOption);
 
     await waitFor(() => {
@@ -66,7 +62,7 @@ describe("ExportReportDropdown", () => {
     });
   });
 
-  it("triggers synchronous Excel export on click", async () => {
+  it("triggers Excel export on click", async () => {
     server.use(
       http.get("*/api/v1/reports/weddings/:uuid/", () => {
         return new HttpResponse("PK\x03\x04 Mock Binary", {
@@ -84,41 +80,11 @@ describe("ExportReportDropdown", () => {
     const button = screen.getByTestId("export-report-dropdown-trigger");
     await user.click(button);
 
-    const excelOption = screen.getByTestId("export-excel-sync");
+    const excelOption = screen.getByTestId("export-excel");
     await user.click(excelOption);
 
     await waitFor(() => {
       expect(window.URL.createObjectURL).toHaveBeenCalled();
-    });
-  });
-
-  it("triggers asynchronous background export on click", async () => {
-    let capturedFormat = "";
-    server.use(
-      http.post("*/api/v1/reports/weddings/:uuid/async/", ({ request }) => {
-        const url = new URL(request.url);
-        capturedFormat = url.searchParams.get("format") || "";
-        return HttpResponse.json(
-          {
-            status: "enqueued",
-            detail: "Geração do relatório iniciada em segundo plano.",
-          },
-          { status: 202 },
-        );
-      }),
-    );
-
-    const user = userEvent.setup();
-    render(<ExportReportDropdown weddingUuid={weddingUuid} />);
-
-    const button = screen.getByTestId("export-report-dropdown-trigger");
-    await user.click(button);
-
-    const asyncPdfOption = screen.getByTestId("export-pdf-async");
-    await user.click(asyncPdfOption);
-
-    await waitFor(() => {
-      expect(capturedFormat).toBe("pdf");
     });
   });
 });
