+++ frontend/src/features/reporting/components/ExportReportDropdown.tsx
@@ -1,11 +1,9 @@
-import { FileDown, FileSpreadsheet, FileText, Loader2, Clock, ChevronDown } from "lucide-react";
+import { FileDown, FileSpreadsheet, FileText, Loader2, ChevronDown } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { useExportReport } from "../hooks/useExportReport";
@@ -25,7 +23,7 @@ export function ExportReportDropdown({
   size = "sm",
   className,
 }: ExportReportDropdownProps) {
-  const { exportSync, exportAsync, isLoading } = useExportReport();
+  const { exportReport, isExporting, exportingFormat } = useExportReport();
 
   const filenamePrefix = weddingName
     ? `relatorio-${weddingName.toLowerCase().replace(/[^a-z0-9]/g, "-")}`
@@ -38,63 +36,36 @@ export function ExportReportDropdown({
           variant={variant}
           size={size}
           className={className}
-          disabled={isLoading}
+          disabled={isExporting}
           data-testid="export-report-dropdown-trigger"
         >
-          {isLoading ? (
+          {isExporting ? (
             <Loader2 className="mr-2 h-4 w-4 animate-spin" />
           ) : (
             <FileDown className="mr-2 h-4 w-4" />
           )}
-          <span>Exportar Relatório</span>
+          <span>{isExporting ? `Exportando ${exportingFormat?.toUpperCase()}...` : "Exportar"}</span>
           <ChevronDown className="ml-1.5 h-3.5 w-3.5 opacity-70" />
         </Button>
       </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="w-56">
-        <DropdownMenuLabel className="text-xs text-muted-foreground uppercase tracking-wider font-semibold">
-          Download Imediato
-        </DropdownMenuLabel>
+      <DropdownMenuContent align="end" className="w-52">
         <DropdownMenuItem
-          onSelect={() => exportSync(weddingUuid, "pdf", filenamePrefix)}
-          onClick={() => exportSync(weddingUuid, "pdf", filenamePrefix)}
+          onSelect={() => exportReport(weddingUuid, "pdf", filenamePrefix)}
+          onClick={() => exportReport(weddingUuid, "pdf", filenamePrefix)}
           className="cursor-pointer"
-          data-testid="export-pdf-sync"
+          data-testid="export-pdf"
         >
-          <FileText className="mr-2 h-4 w-4 text-rose-500" />
-          <span>Relatório PDF (.pdf)</span>
+          <FileText className="mr-2 h-4 w-4 text-purple-600" />
+          <span>Relatório (PDF)</span>
         </DropdownMenuItem>
         <DropdownMenuItem
-          onSelect={() => exportSync(weddingUuid, "excel", filenamePrefix)}
-          onClick={() => exportSync(weddingUuid, "excel", filenamePrefix)}
+          onSelect={() => exportReport(weddingUuid, "excel", filenamePrefix)}
+          onClick={() => exportReport(weddingUuid, "excel", filenamePrefix)}
           className="cursor-pointer"
-          data-testid="export-excel-sync"
+          data-testid="export-excel"
         >
           <FileSpreadsheet className="mr-2 h-4 w-4 text-emerald-600" />
-          <span>Planilha Excel (.xlsx)</span>
-        </DropdownMenuItem>
-
-        <DropdownMenuSeparator />
-
-        <DropdownMenuLabel className="text-xs text-muted-foreground uppercase tracking-wider font-semibold">
-          Segundo Plano (Worker)
-        </DropdownMenuLabel>
-        <DropdownMenuItem
-          onSelect={() => exportAsync(weddingUuid, "pdf")}
-          onClick={() => exportAsync(weddingUuid, "pdf")}
-          className="cursor-pointer"
-          data-testid="export-pdf-async"
-        >
-          <Clock className="mr-2 h-4 w-4 text-muted-foreground" />
-          <span>Processar PDF (Notificar)</span>
-        </DropdownMenuItem>
-        <DropdownMenuItem
-          onSelect={() => exportAsync(weddingUuid, "excel")}
-          onClick={() => exportAsync(weddingUuid, "excel")}
-          className="cursor-pointer"
-          data-testid="export-excel-async"
-        >
-          <Clock className="mr-2 h-4 w-4 text-muted-foreground" />
-          <span>Processar Excel (Notificar)</span>
+          <span>Planilha (Excel)</span>
         </DropdownMenuItem>
       </DropdownMenuContent>
     </DropdownMenu>
